+++ examples/static-assets/src/index.ts
@@ -1,0 +1,3 @@
+import './style/index.scss';
+
+document.body.innerHTML = '<div id="root"><div>Hello World</div></div>';